+++ test/test193.js
@@ -8,10 +8,19 @@ if(typeof exports === 'object') {
 //if(typeof exports != 'object') {
 
 describe('Test 193 - GROUP BY formula', function() {
+
+    console.log(alasql.parse('SELECT a FROM ? GROUP BY a % 2').toString());
+
     it("1. GROUP BY column", function(done) {
         var data = [{a:1},{a:1},{a:2},{a:3},{a:1},{a:2}];
-        var res = alasql('SELECT a FROM ? ORDER BY 1-a',[data]);
-        assert.deepEqual(res,[ { a: 3 }, { a: 2 }, { a: 2 }, { a: 1 }, { a: 1 }, { a: 1 } ]);
+        var res = alasql('SELECT a FROM ? GROUP BY a ',[data]);
+        console.log(res);
     	done();
+    });
+    it("2. GROUP BY formula", function(done) {
+        var data = [{a:1},{a:1},{a:2},{a:3},{a:1},{a:2}];
+        var res = alasql('SELECT a FROM ? GROUP BY a % 2',[data]);
+        console.log(res);
+        done();
     });
 });